cleaned up the folder and uploaded different trained models
