--- python_app/Test_SoundGeneratorVAE_V2/model/INFO training blues.txt
@@ -1,3 +1,0 @@
-Trained with blues files
-Batchsize = 2
-Epochs = 20
